Product Page update and Bug fix on Account Page
Added two more items to the product page to highlight the php code determining if a product is in stock or not. Also fixed a bug on the account page where if a user updated there details, went to the home page and back to their account, all details would be missing.